feat: Comunicação entre microservicos de Venda e Veiculo

+++ src/main/kotlin/com/curso/dto/input/VendaInput.kt
@@ -4,7 +4,7 @@ import java.math.BigDecimal
 
 data class VendaInput(
   val cliente : String,
-  val veiculo : Int,
+  val veiculo : Long,
   val valor : BigDecimal,
   val quantidadeParcelas : Int
 )
